normi para signals
